requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pyyaml==6.0.2
cryptography==43.0.1
python-multipart==0.0.12

